fix: use gamma = 1.0 gets completely different result
The old gamma value was 0.99, which is pretty much the default for other
RL demo found online. But I guess for financial trading environment,
this is some how related to time value of money, which may be very
different from the meaning when playing games. The true behavior is yet
to be revealed.